Extract imprint to new file
